immuno_db_app/requirements.txt
django==4.0.6
django-crispy-forms==1.14.0
djangorestframework==3.13.1
django-tables2==2.4.1
numpy==1.23.3
pandas==1.5.0
plotly==5.10.0
mypy==0.971
psycopg2==2.9.3
xlsxwriter==3.0.3
gunicorn==20.1.0
django-filter==22.1
django-bootstrap5==21.3
django-select2==8.1.2
tablib==3.5.0
matplotlib==3.5.3
adjustText==0.8
